1.集成minio 2.删除原来导入模板 3.军事训练导入特殊处理 4.新增项目负责人设置

+++ stuup_common/src/main/java/com/poho/stuup/model/vo/ClassRankVO.java
@@ -2,6 +2,8 @@ package com.poho.stuup.model.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author BUNGA
  * @description: 全校学生排名对象
@@ -43,7 +45,7 @@ public class ClassRankVO {
     /**
      * 成长值
      */
-    private Integer score;
+    private BigDecimal score;
 
 
 }
