+++ stuup_common/src/main/java/com/poho/stuup/model/vo/MajorRankVO.java
@@ -2,6 +2,8 @@ package com.poho.stuup.model.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @author BUNGA
  * @description: 全校专业排名对象
@@ -28,7 +30,7 @@ public class MajorRankVO {
     /**
      * 成长值
      */
-    private int score;
+    private BigDecimal score;
 
 
 }
